requirements.txt
websocket-client>=1.0.0
numpy>=1.20.0
sounddevice>=0.4.4
python-dotenv>=0.19.0
colorama>=0.4.4
pynput>=1.7.6
pyperclip>=1.8.2 